Act Professional Guys Read my Message!
NO - Don't Ever Put Testtttt, This website is in full use. we use this to get out to sponsors. It Is in beta Because we are using it!!!!  And stop calling each other stupid and dumb!!! WE ARE IN HIGH SCHOOL!!!!! The Maps Looks good, and is there a reason we have the twitter api back? is it not a conflict with the rules? Hope You Two Are Messing Around. Btw caterpillar is Interested in sponsoring... We need this site to look good.

Thank You
 - Josiah Papovitch